+++ lib/Ansifmt.ml
@@ -1,5 +1,10 @@
 module Color = Color
 module IO = IO
 module Formatting = Formatting
-module Prelude = Prelude
 module Styling = Styling
+
+type color = Color.t
+type styling = Styling.t
+type stylizer = Formatting.Stylizer.t
+
+let format = Formatting.Util.format