+++ lib/IO.ml
@@ -1,8 +1,14 @@
+open Formatting
+
 (** [print_formatted ?stylizer ?line_end ?out value ~using]
-    prints [value] to [out] by formatting it with [using] -
-    which provides tokenization of [value] - and [stylizer]
-    which renders the resulting tokens into a pretty-printable
-    string. [line_end] is appended at the end of the string.
+    prints [value] to [out] by formatting it with [using] --
+    which provides conversion of [value] to a formatting element
+    -- and [stylizer] which renders the resulting element into
+    a pretty-printable string. [line_end] is appended at the end
+    of the string.
+
+    This function is meant to have an interface resembling
+    Python's [print] function.
     
     Defaults if not provided:
       - [stylizer]: the default stylizer provided by [ansifmt]
@@ -10,18 +16,13 @@
       - [out]: the standard output ([stdout]) *)
 let print_formatted
   : type t.
-    ?stylizer:Formatting.Stylizer.t
-    -> ?parentheses:string * string
+    ?stylizer:Stylizer.t
     -> ?line_end:string
     -> ?out:out_channel
     -> t
-    -> using:(module Formatting.TOKENIZABLE with type t = t)
+    -> using:(module Interfaces.TO_ELEMENT with type t = t)
     -> unit
   =
-  fun ?stylizer ?parentheses ?(line_end = "\n") ?(out = stdout) value ~using:(module M) ->
-  Printf.fprintf
-    out
-    "%s%s"
-    (Formatting.Util.format ?stylizer ?parentheses value ~using:(module M))
-    line_end
+  fun ?stylizer ?(line_end = "\n") ?(out = stdout) value ~using:(module M) ->
+  Printf.fprintf out "%s%s" (Util.format ?stylizer value ~using:(module M)) line_end
 ;;